disorganized mess

+++ tests/functional/test_streamify.py
@@ -1,4 +1,3 @@
-import logging
 import os
 import shutil
 
@@ -7,7 +6,6 @@ from twisted.trial.unittest import TestCase
 from twisted.internet import defer, threads
 
 from lbrynet import conf
-from lbrynet.lbryfile.EncryptedFileMetadataManager import EncryptedFileMetadataManager
 from lbrynet.lbryfile.EncryptedFileMetadataManager import DBEncryptedFileMetadataManager
 from lbrynet.lbryfilemanager.EncryptedFileManager import EncryptedFileManager
 from lbrynet.core.Session import Session
@@ -20,8 +18,6 @@ from lbrynet.core.PeerManager import PeerManager
 from lbrynet.core.RateLimiter import DummyRateLimiter, RateLimiter
 
 from tests import mocks
-
-logging.getLogger("lbrynet").setLevel(logging.CRITICAL)
 
 FakeNode = mocks.Node
 FakeWallet = mocks.Wallet
@@ -51,9 +47,10 @@ class TestStreamify(TestCase):
             d.addCallback(lambda _: self.stream_info_manager.stop())
 
         def delete_test_env():
-            shutil.rmtree('client')
-            if os.path.exists("test_file"):
-                os.remove("test_file")
+            if os.path.isdir('client'):
+                shutil.rmtree('client')
+            if os.path.exists('test_file'):
+                os.remove('test_file')
 
         d.addCallback(lambda _: threads.deferToThread(delete_test_env))
         return d
@@ -66,22 +63,21 @@ class TestStreamify(TestCase):
         rate_limiter = DummyRateLimiter()
         sd_identifier = StreamDescriptorIdentifier()
 
-
         db_dir = "client"
         blob_dir = os.path.join(db_dir, "blobfiles")
         os.mkdir(db_dir)
         os.mkdir(blob_dir)
 
         self.session = Session(
-            conf.ADJUSTABLE_SETTINGS['data_rate'][1], db_dir=db_dir, lbryid="abcd",
-            peer_finder=peer_finder, hash_announcer=hash_announcer,
+            conf.ADJUSTABLE_SETTINGS['data_rate'][1], db_dir=db_dir,
+            lbryid="abcd", peer_finder=peer_finder, hash_announcer=hash_announcer,
             blob_dir=blob_dir, peer_port=5553,
             use_upnp=False, rate_limiter=rate_limiter, wallet=wallet,
             blob_tracker_class=DummyBlobAvailabilityTracker,
             is_generous=self.is_generous
         )
 
-        self.stream_info_manager = EncryptedFileMetadataManager()
+        self.stream_info_manager = DBEncryptedFileMetadataManager(self.session.storage)
 
         self.lbry_file_manager = EncryptedFileManager(
             self.session, self.stream_info_manager, sd_identifier)
@@ -113,6 +109,7 @@ class TestStreamify(TestCase):
 
         d.addCallback(lambda _: create_stream())
         d.addCallback(verify_stream_descriptor_file)
+        d.addCallback(lambda _: self.take_down_env())
         return d
 
     def test_create_and_combine_stream(self):
@@ -136,29 +133,10 @@ class TestStreamify(TestCase):
             blob_tracker_class=DummyBlobAvailabilityTracker
         )
 
-        self.stream_info_manager = DBEncryptedFileMetadataManager(self.session.db_dir)
+        self.stream_info_manager = DBEncryptedFileMetadataManager(self.session.storage)
 
         self.lbry_file_manager = EncryptedFileManager(
             self.session, self.stream_info_manager, sd_identifier)
-
-        def start_lbry_file(lbry_file):
-            logging.debug("Calling lbry_file.start()")
-            d = lbry_file.start()
-            return d
-
-        def combine_stream(stream_hash):
-            prm = self.session.payment_rate_manager
-            d = self.lbry_file_manager.add_lbry_file(stream_hash, prm)
-            d.addCallback(start_lbry_file)
-
-            def check_md5_sum():
-                f = open('test_file')
-                hashsum = MD5.new()
-                hashsum.update(f.read())
-                self.assertEqual(hashsum.hexdigest(), "68959747edc73df45e45db6379dd7b3b")
-
-            d.addCallback(lambda _: check_md5_sum())
-            return d
 
         @defer.inlineCallbacks
         def create_stream():
@@ -166,12 +144,19 @@ class TestStreamify(TestCase):
             stream_hash = yield create_lbry_file(self.session, self.lbry_file_manager, "test_file",
                                              test_file, suggested_file_name="test_file")
             yield publish_sd_blob(self.stream_info_manager, self.session.blob_manager, stream_hash)
-            defer.returnValue(stream_hash)
+
+            prm = self.session.payment_rate_manager
+            lbry_file = yield self.lbry_file_manager.add_lbry_file(stream_hash, prm)
+            yield lbry_file.start()
+            f = open('test_file')
+            hashsum = MD5.new()
+            hashsum.update(f.read())
+            self.assertEqual(hashsum.hexdigest(), "68959747edc73df45e45db6379dd7b3b")
+            defer.returnValue(None)
 
         d = self.session.setup()
         d.addCallback(lambda _: self.stream_info_manager.setup())
         d.addCallback(lambda _: add_lbry_file_to_sd_identifier(sd_identifier))
         d.addCallback(lambda _: self.lbry_file_manager.setup())
         d.addCallback(lambda _: create_stream())
-        d.addCallback(combine_stream)
         return d
